🚚 rename useToaster to index
