+++ frontend/src/components/Toaster/index.tsx
@@ -1,4 +1,4 @@
-import { useToaster } from "src/hooks/useToaster/useToaster";
+import { useToaster } from "src/hooks/useToaster";
 import View from "./View";
 
 export const Toaster = () => {
